refactor: hide exception class
