+++ contrib/simple_regression/simple_regression.py
@@ -1,5 +1,6 @@
 import numpy as np
-from cofi_espresso import EspressoProblem, InvalidExampleError
+from cofi_espresso import EspressoProblem
+from cofi_espresso.exceptions import InvalidExampleError
 
 
 class SimpleRegression(EspressoProblem):
