+++ contrib/xray_tomography/xray_tomography.py
@@ -1,5 +1,6 @@
 import numpy as np
-from cofi_espresso import EspressoProblem, InvalidExampleError
+from cofi_espresso import EspressoProblem
+from cofi_espresso.exceptions import InvalidExampleError
 from cofi_espresso.utils import loadtxt, absolute_path
 from PIL import Image
 import matplotlib.pyplot as plt
